+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The changelog lists relevant feature changes between each release. Search GitHub issues and pull requests for smaller issues.
 
+## [unreleased]
+- add converter for `gbfs.prod.sharedmobility.ch/v2/gbfs/pickebike_basel` feed: set valid pricing_plan_id
+
 ## 2024-09-17
 - add converter for `apis.deutschebahn.com` feeds: remove capacity attribute
 - change converter for `gbfs.nextbike.net` feeds: set `propulsion_type=human` if not defined